Se cambio la ubicacion del enum ahora esta en un archivo propio
