+++ src/hoja-trabajo/dto/estado-viaje.enum.ts
@@ -1,0 +1,7 @@
+export enum EstadoHojaTrabajo {
+    PROGRAMADO = 'programado',
+    EN_CURSO = 'en curso',
+    FINALIZADO = 'finalizado',
+    SUSPENDIDO = 'suspendido',
+    CANCELADO = 'cancelado'
+  }
